+++ jackson/src/main/scala/com/fasterxml/jackson/databind/ObjectMapperCopier.scala
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.module.scala.experimental.ScalaObjectMapper
 object ObjectMapperCopier {
 
   // Workaround since calling objectMapper.copy on "ObjectMapper with ScalaObjectMapper" fails the _checkInvalidCopy check
-  def copy(objectMapper: ObjectMapper) = {
+  def copy(objectMapper: ObjectMapper): ObjectMapper with ScalaObjectMapper = {
     new ObjectMapper(objectMapper) with ScalaObjectMapper
   }
 }