+++ jackson/src/main/scala/com/twitter/finatra/json/internal/caseclass/exceptions/FinatraJsonMappingException.scala
@@ -5,6 +5,11 @@ import com.fasterxml.jackson.databind.JsonMappingException
 /**
  * Exception for handling Finatra-specific errors that may otherwise be valid
  * in the eyes of Jackson.
- * @param msg gets presented to the end user, so don't leak implementation details.
+ *
+ * @note the given [[msg]] generally passes all the way up the stack to the caller of
+ *       the method throwing this Exception, thus care should be taken to not leak any
+ *       internal implementation details.
+ *
+ * @param msg exception message.
  */
-class FinatraJsonMappingException(msg: String) extends JsonMappingException(msg) {}
+class FinatraJsonMappingException(msg: String) extends JsonMappingException(null, msg)