+++ jackson/src/main/scala/com/twitter/finatra/json/internal/serde/DurationStringSerializer.scala
@@ -5,9 +5,14 @@ import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.databind.ser.std.StdSerializer
 import com.twitter.util.Duration
 
-private[finatra] object TwitterUtilDurationStringSerializer
+private[finatra] object DurationStringSerializer
     extends StdSerializer[Duration](classOf[Duration]) {
-  override def serialize(value: Duration, jgen: JsonGenerator, provider: SerializerProvider): Unit = {
+
+  override def serialize(
+    value: Duration,
+    jgen: JsonGenerator,
+    provider: SerializerProvider
+  ): Unit = {
     jgen.writeString(value.toString)
   }
 }