+++ jackson/src/main/scala/com/twitter/finatra/json/internal/serde/TimeStringSerializer.scala
@@ -5,8 +5,13 @@ import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.databind.ser.std.StdSerializer
 import com.twitter.util.Time
 
-private[finatra] object TwitterUtilTimeStringSerializer extends StdSerializer[Time](classOf[Time]) {
-  override def serialize(value: Time, jgen: JsonGenerator, provider: SerializerProvider): Unit = {
+private[finatra] object TimeStringSerializer extends StdSerializer[Time](classOf[Time]) {
+
+  override def serialize(
+    value: Time,
+    jgen: JsonGenerator,
+    provider: SerializerProvider
+  ): Unit = {
     jgen.writeString(value.toString)
   }
 }