+++ jackson/src/test/scala/com/twitter/finatra/json/tests/internal/JodaTimeModuleTest.scala
@@ -3,11 +3,11 @@ package com.twitter.finatra.json.tests.internal
 import com.fasterxml.jackson.databind.{ObjectMapper, SerializationFeature}
 import com.fasterxml.jackson.datatype.joda.JodaModule
 import com.twitter.finatra.json.FinatraObjectMapper
-import com.twitter.finatra.json.internal.serde.FinatraSerDeSimpleModule
+import com.twitter.finatra.json.internal.serde.SerDeSimpleModule
 import com.twitter.inject.Test
 import org.joda.time.{DateTime, DateTimeZone}
 
-class FinatraJodaTimeModuleTest extends Test {
+class JodaTimeModuleTest extends Test {
 
   val nowUtc = DateTime.now.withZone(DateTimeZone.UTC)
 
@@ -27,20 +27,20 @@ class FinatraJodaTimeModuleTest extends Test {
 
   test("deserialize text") {
     val mapper = new ObjectMapper()
-    mapper.registerModule(FinatraSerDeSimpleModule)
+    mapper.registerModule(SerDeSimpleModule)
     mapper.readValue(quote(nowUtc.toString()), classOf[DateTime]) should equal(nowUtc)
   }
 
   test("deserialize text with FinatraObjectMapper") {
     val mapper = FinatraObjectMapper.create()
-    mapper.registerModule(FinatraSerDeSimpleModule)
+    mapper.registerModule(SerDeSimpleModule)
     mapper.parse[DateTime](quote(nowUtc.toString())) should equal(nowUtc)
   }
 
   test("deserialize long") {
     val mapper = new ObjectMapper()
     mapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, true)
-    mapper.registerModule(FinatraSerDeSimpleModule)
+    mapper.registerModule(SerDeSimpleModule)
     mapper.readValue(nowUtc.getMillis.toString(), classOf[DateTime]) should equal(nowUtc)
   }