+++ jackson/src/test/scala/com/twitter/finatra/validation/InvalidValidator.scala
@@ -1,10 +1,12 @@
 package com.twitter.finatra.validation
 
+import scala.util.control.NoStackTrace
+
 class InvalidValidator(
   validationMessageResolver: ValidationMessageResolver,
   annotation: InvalidValidationInternal
 ) extends Validator[Any, Any](validationMessageResolver, annotation) {
 
   override def isValid(value: Any): ValidationResult =
-    throw new RuntimeException("validator foo error")
+    throw new RuntimeException("validator foo error") with NoStackTrace
 }